+++ lib/src/preference_event.dart
@@ -4,7 +4,7 @@ abstract class PreferenceEvent extends Equatable {
   PreferenceEvent([List props = const []]) : super(props);
 }
 
-class FetchPreference extends PreferenceEvent {
+class UpdatePreference extends PreferenceEvent {
   @override
-  String toString() => 'FetchPreference';
+  String toString() => 'UpdatePreference';
 }