+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4
+
+- Change bloc init state to loaded preference state
+- Remove PreferenceLoading, PreferenceLoaded state
+- Rename FetchPreference event to UpdatePreference
+
 ## 0.0.3
 
 - Add documentation comments